+++ src/lib/AudioDecoder.ts
@@ -39,7 +39,8 @@ export class AudioDecoder extends Transform implements MediaStream {
       this.busy = false;
       this.emit('ready');
       callback();
-    })().catch(callback);
+    })()
+      .catch(callback);
   }
 
   _transform(packet: any, encoding: BufferEncoding, callback: TransformCallback): void {
@@ -56,7 +57,8 @@ export class AudioDecoder extends Transform implements MediaStream {
       }
       this.busy = false;
       callback();
-    })().catch(callback);
+    })()
+      .catch(callback);
   }
 
   coder() {